Updated script to check that branch change succeeded

+++ gh-deploy.ps1
@@ -1,12 +1,15 @@
 Try
 {
     $current_branch = git rev-parse --abbrev-ref HEAD
-    git checkout gh-deploy
-    gulp
-    git add dist -f
-    git commit -m "deploying..."
-    git subtree push --prefix dist origin gh-pages
-    Remove-Item dist
+    $ckerr = git checkout gh-deploy
+    if ($?)
+    {
+        gulp
+        git add dist -f
+        git commit -m "deploying..."
+        git subtree push --prefix dist origin gh-pages
+        Remove-Item dist
+    }
 }
 Finally
 {
